refactor: 클래스명을 MessageServiceTest로 변경 및 시간에 따른 리마인드 테스트 이동
